[WIP]: Simple authentication implementation. (#7)
* Began implementing authentication

* Properly use Dropwizard's authentication systems, and store API keys in
the database.

To be able to protect a specific route with authentication, annotate the
path with @RolesAllowed(AuthRoles.VOTER | AuthRoles.REPRESENTATIVE).

* Move raw sql strings to a file.

* Cleaned up bits of the representation to match the SQL.

* Added documentation for API

* Fixed markdown typo

+++ sql/create_district_table.sql
@@ -1,5 +1,5 @@
 create table if not exists district
    (id bigint primary key auto_increment,
     state varchar(2) not null,
-    district_num int not null,
+    district_num bigint not null,
     is_senate boolean);
